% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_create.R
\name{app_create}
\alias{app_create}
\title{Creates ABIC Shiny App}
\usage{
app_create(path, app_name = basename(path), open = NULL)
}
\arguments{
\item{path}{Name of folder to create app}

\item{app_name}{Package name to use. Defaults to basename(path)}

\item{open}{Open RStudio Project?}
}
\description{
Creates ABIC Shiny App
}
