% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_create.R
\name{app_create}
\alias{app_create}
\title{Creates DynastyProcess Shiny App}
\usage{
app_create(path, app_name = basename(path), install_deps = NULL, open = NULL)
}
\arguments{
\item{path}{Name of folder to create app}

\item{app_name}{Package name to use. Defaults to basename(path)}

\item{open}{Logical: open RStudio Project?}

\item{install_dependencies}{Logical: install packages?}
}
\description{
Creates DynastyProcess Shiny App
}
